Configure build script
